53: Remove 'AutoCorrect' automatical option
